moving histogram classes to another package
